requirements.txt
Django==2.0.2
django-pure-pagination==0.3.0
Markdown==2.6.11
mysqlclient==1.3.12
Pygments==2.2.0
pytz==2017.3
